+++ e2e/fixtures/citizen/createCase/C100/otherProceedings/occupationOrderDetailsContent.ts
