+++ e2e/pages/citizen/createCase/C100/otherProceedings/forcedMarriageProtectionOrderDetailsPage.ts
@@ -4,7 +4,7 @@ import { Helpers } from "../../../../../common/helpers";
 import AccessibilityTestHelper from "../../../../../common/accessibilityTestHelper";
 import { Page } from "@playwright/test";
 import { CitizensOtherProceedingsUniqueSelectors } from "../../../../../common/commonUniqueSelectors";
-import { ForcedMarriageProtectionOrderDetailsContent } from "../../../../../fixtures/citizen/createCase/C100/otherProceedings/forcedMarriageProtection/forcedMarriageProtectionOrderDetailsContent";
+import { ForcedMarriageProtectionOrderDetailsContent } from "../../../../../fixtures/citizen/createCase/C100/otherProceedings/forcedMarriageProtectionOrderDetailsContent";
 
 interface ForcedMarriageProtectionOrderDetailsPageOptions {
   page: Page;
